add x86_64 build target alongside existing i386
Make the loader dual-arch: `make ARCH=i386` keeps the existing 32-bit
behavior, `make ARCH=x86_64` produces a 64-bit nloader that can map and
run PE32+ x86-64 native binaries on Linux.

Highlights:

- common.mak/Makefile: ARCH-conditional CFLAGS/YASM/objcopy. Default
  THREADED on with -pthread for both arches.
- nt_structs.h: PE32+ optional header struct, arch-conditional
  PE_MACHINE_EXPECTED/PE_OPT_MAGIC/PE_ORDINAL_FLAG/pe_thunk_t. WINAPI
  expands to __attribute__((ms_abi)) on x86_64 so libntdll receives args
  in the Win64 register set the loaded PE emits at every call site.
  FORWARD_FUNCTION on x86_64 emits a Win64 -> SysV register shuffle
  before the tail-jmp into libc.
- nt_compat.h: set_teb_register() unifies modify_ldt (i386) and
  arch_prctl(ARCH_SET_GS) (x86_64); SharedUserData base remains
  0x7ffe0000.
- loader.c: PE parser uses pe_image_optional_hdr_t, pe_thunk_t, the
  arch-conditional ordinal mask. ImageBase no longer truncated to 32
  bits. TEB mmap sized via PESALIGN(sizeof(TEB), pagesize) since the
  Win64 TEB doesn't fit a single page. peb/ldr_data are file-static so
  their addresses outlive setup_nloader and survive the TEB pointer
  hand-off. Stub trampoline rewritten to a 17-byte
  push-imm32/mov-rax-imm64/jmp-rax sequence (rel32 jmp can't reach the
  dispatcher when stubs are mmap'd far from the PIE base). Missing
  import .so falls through to stubs instead of aborting.
- stubs_x64.asm: new dispatcher matching the runtime stub layout. Tail-
  call teardown maintains rsp%16==0 before the resolved-fn jmp so it
  sees Win64-ABI entry alignment. to_ep switches to the TEB stack with
  rsp%16==0 prior to `call`. LogModule path uses SysV registers since
  LogModule is now SysV (its va_list goes into libc vfprintf).
- libs/ntdll/_chkstk.asm: x86_64 implementation that probes pages on
  RAX bytes without touching RSP, gs:0x10/0x08 for stack limit/base.
- libs/ntdll/sync.c: rewrite around a single state mutex per event.
  Fixes the `(ObjectCount = 2)` typo in NtWaitForMultipleObjects, the
  cond_mutex deadlock in the already-signaled path, and gives
  WaitForMultipleObjects a real implementation (sequential WaitAll,
  10ms-poll WaitAny). Drops the dead non-THREADED branch.
- libs/ntdll/thread.c: pthread-only path. Critical sections always use
  recursive pthread_mutex_t. Per-thread TEB created and installed via
  install_teb() so each pthread sees its own NtCurrentTeb(). Drop the
  setjmp/longjmp fallback and the abort() in NtTerminateThread for
  non-zero exit codes.
- libs/ntdll/{bitmap,heap,io,ioctl,mm,scancodes}.c: small adjustments
  for size_t/uintptr_t and PE32+ paths.

+++ libs/ntdll/ioctl.c
@@ -520,11 +520,11 @@ NTSTATUS NTAPI NtDeviceIoControlFile(HANDLE FileHandle, HANDLE Event, PIO_APC_RO
 
             if (geo->Cylinders.QuadPart < 2)
             {
-                fprintf(stderr, "ntdll.NtDeviceIoControlFile(IOCTL_DISK_GET_DRIVE_GEOMETRY, Cylinders = %llu) - TOO SMALL!!\n", geo->Cylinders.QuadPart);
+                fprintf(stderr, "ntdll.NtDeviceIoControlFile(IOCTL_DISK_GET_DRIVE_GEOMETRY, Cylinders = %llu) - TOO SMALL!!\n", (unsigned long long)geo->Cylinders.QuadPart);
                 return (IoStatusBlock->u.Status = STATUS_INVALID_PARAMETER);
             }
 
-            Log("ntdll.NtDeviceIoControlFile(IOCTL_DISK_GET_DRIVE_GEOMETRY, Cylinders = %llu)\n", geo->Cylinders.QuadPart);
+            Log("ntdll.NtDeviceIoControlFile(IOCTL_DISK_GET_DRIVE_GEOMETRY, Cylinders = %llu)\n", (unsigned long long)geo->Cylinders.QuadPart);
 
             IoStatusBlock->Information = sizeof(DISK_GEOMETRY);
             IoStatusBlock->u.Status = STATUS_SUCCESS;
